test(BA-7150): drop the merged-read resolver wiring spies
Both resolvers pass their arguments straight to an adapter method, so the
tests asserted the call wiring rather than a contract. Keep the type
mapping test, under the name the sibling GQL type tests use.
