--- tests/unit/manager/api/gql/app_config/BUILD
@@ -1,3 +1,0 @@
-python_tests(
-    name="tests",
-)
